Cleanup and bash-ify .bashrc
The sourcing logic can be unified, and using an explicit `source` rather
than a `.` is more clear.

As well, vimglob can be implemented using pure bash, which performs
better and is less error prone, considering the current lack of quoting.
We implement the function with a subshell ( ) instead of a grouping { }
so that the shopt change doesn't affect our current context, and then we
exec vim instead of the ordinary invocation, so that despite the
subshell, we don't actually incur two forks, so the cost is the same
between subshell and grouping.

The [[ ]] operator in bash automatically quotes one operand, so we use
that instead.

The PATH addition now prefers an explicit $HOME to ~, and it also
prioritizes utilities in $HOME/bin over system ones.